requirements.txt
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
streamlit>=1.29.0
bcrypt>=4.0.1
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
boto3>=1.34.0 